FO-2897: Legg til custom lenke wrapper
Så flyten kan funke rett for veiledere også
